feat(occurrence/[occurrence_id]/page.tsx): add generateMetadata function to set the page title to the occurrence message fix(components/BookmarksTable.tsx): fix occurrence link href to use occurrence_id instead of occurrence.id fix(lib/actions/occurrenceActions.ts): add error handling for session not found when creating or removing occurrence bookmarks

+++ lib/actions/occurrenceActions.ts
@@ -53,30 +53,36 @@ export async function performReplay(context: Context): Promise<string> {
 // Function to create a bookmark for a user
 export async function createOccurrenceBookmark(occurrenceId: string) {
   const session = await getServerSession();
+  if (!session) {
+    throw new Error('Session not found');
+  }
 
   await prisma.occurrenceBookmark.create({
     data: {
-      user_id: session?.user?.id!,
+      user_id: session.user?.id!,
       occurrence_id: occurrenceId,
     },
   });
 
-  revalidatePath('/bookmarks')
   revalidatePath(`/occurrences/${occurrenceId}`)
+  revalidatePath('/bookmarks')
 }
 
 export async function removeOccurrenceBookmark(occurrenceId: string) {
   const session = await getServerSession();
+  if (!session) {
+    throw new Error('Session not found');
+  }
 
   await prisma.occurrenceBookmark.delete({
     where: {
       user_id_occurrence_id: {
-        user_id: session?.user?.id!,
+        user_id: session.user?.id!,
         occurrence_id: occurrenceId,
       },
     },
   });
 
-  revalidatePath('/bookmarks')
   revalidatePath(`/occurrences/${occurrenceId}`)
+  revalidatePath('/bookmarks')
 }
